requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
fastapi
uvicorn
watchdog
langgraph-cli[inmem]
python-dotenv
streamlit>=1.32.0
requests
